ensure correct binary is used and delete error
As learned during the last Regen Network testnet: it makes sense for the command line and cosmovisor to use the same binary for `osmosisd`. This means that the binary should be moved, not copied, to the cosmovisor directory, and the cosmovisor directory should be added to the PATH. If you don't do it this way, then you risk having two different binaries on the system, and those binaries eventually being different versions (after network upgrades), and weird things happening.

Also, deleted an extraneous line about adding a new `.osmosisd` directory -- that's from the `init` section.